rag_chatbot/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9  # For Neon PostgreSQL
qdrant-client==1.7.0
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.8
sentence-transformers==2.2.2 # For embeddings
# For testing
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
